third_party/blink/web_tests/external/wpt/trusted-types/TrustedTypePolicyFactory-getAttributeType-expected.txt

This is a testharness.js-based test.
[FAIL] getAttributeType(\n          "script",\n          "href",\n          "http://www.w3.org/2000/svg",\n          "null") == "TrustedScriptURL"
  assert_equals: expected (string) "TrustedScriptURL" but got (object) null
[FAIL] getAttributeType(\n          "script",\n          "href",\n          "http://www.w3.org/2000/svg",\n          "http://www.w3.org/1999/xlink") == "TrustedScriptURL"
  assert_equals: expected (string) "TrustedScriptURL" but got (object) null
[FAIL] getAttributeType(\n          "foo",\n          "onmouseup",\n          "https://example.com/namespace",\n          "null") == "null"
  assert_equals: expected (object) null but got (string) "TrustedScript"
Harness: the test ran to completion.
